+++ dflib/src/test/java/org/dflib/exp/flow/IfNullTest.java
@@ -8,7 +8,20 @@ import org.junit.jupiter.api.Test;
 
 import static org.dflib.Exp.*;
 
-public class IfNullExpTest extends BaseExpTest {
+public class IfNullTest extends BaseExpTest {
+
+    @Test
+    public void col() {
+        Exp<?> noNulls = ifNull($col("a"), $col("b"));
+
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
+                "1", "2",
+                null, "5",
+                null, "6",
+                null, null);
+
+        new SeriesAsserts(noNulls.eval(df)).expectData("1", "5", "6", null);
+    }
 
     @Test
     public void string() {
@@ -37,19 +50,6 @@ public class IfNullExpTest extends BaseExpTest {
     }
 
     @Test
-    public void val() {
-        Exp<?> noNulls = ifNull($int("a"), 77);
-
-        DataFrame df = DataFrame.foldByRow("a").of(
-                1,
-                null,
-                8,
-                null);
-
-        new SeriesAsserts(noNulls.eval(df)).expectData(1, 77, 8, 77);
-    }
-
-    @Test
     public void testEquals() {
         assertExpEquals(
                 ifNull($int("a"), $int(1)),
@@ -74,4 +74,5 @@ public class IfNullExpTest extends BaseExpTest {
                 ifNull($int("a"), $int(1)),
                 ifNull($int("a"), $int(2)));
     }
+
 }